requirements.txt
openai==2.8.1
qdrant-client==1.16.1
python-dotenv==1.2.1
